clean up more game_error stragglers missed in #2918

+++ lib/engine/step/g_1828/remove_tokens.rb
@@ -50,7 +50,7 @@ module Engine
           city = action.city
           token = city.tokens[action.slot]
           hex = city.hex
-          @game.game_error("Cannot remove #{token.corporation.name} token") unless available_hex(entity, hex)
+          raise GameError, "Cannot remove #{token.corporation.name} token" unless available_hex(entity, hex)
 
           @log << "#{entity.name} removes token from #{hex.name} (#{hex.location_name})"
           token.destroy!
